requirements.txt
et-xmlfile==1.1.0
lxml==4.9.3
openpyxl==3.1.2
Pillow==10.1.0
python-docx==1.1.0
python-pptx==0.6.23
setuptools==68.2.2
typing_extensions==4.8.0
wheel==0.41.2
XlsxWriter==3.1.9
